+++ src/compiler/index.ts
@@ -1,4 +1,4 @@
-export { KipperCompiler } from "./compiler";
-export { KipperErrorListener } from "./error-handler";
+export * from "./compiler";
+export * from "./error-handler";
 export * as parser from "./parser";
-export { ParserFile } from "./parser-file";
+export * from "./parser-file";